fix: resets DeviceBrightness on inactive screen
